Create Database schema for basic session tracking
Create a new session and slot table, and modify the task table to hold a
foreign key of what session its for.

To create a new session, we create a row in the Session table with a
start and end time. Then, we can add all the tasks the user wants to do
to the tasks table with that session. Once we generate slots, we'll add
them at the same time.

This will break the current tasks UI.