Make slot height non-proportional to window height
Rather than express the height of each slot as a percentage of the
window height, which leads to very squished ui on mobile, have it be a
constant amount of ems.

This makes the UI usable when a phone is in landscape mode.

The scaling is logorithmic to avoid short sessions being dwarfed by
longer work sessions. To avoid issues, we shouldn't schedule breaks or
sessions shorter than four minutes.