Implement basic session restoring
Add an extra field to the Session schema to store the intended duration
of the session. We set this once during session creation, and don't
shorten it if the session ends early, like for end point.

If the user chooses to restore a previous session (WE NEED A BETTER WORD
FOR THIS THAN RESTORE), we make a new session of the same duration as
their last session.

We also slightly modify how the API works. current_session doesn't check
time, it just gives back whatever session the user is doing, or did
last, but /api/session/active has been changed to check that the current
end time hasn't passed yet.