Include PostSessionFeedback modal in CurrSessView
Create a PostSessionFeedback component, and button that brings it up on
screen, for each work session in the curr.

We make the containing div take up the whole session box, so that with
some margin auto magic, we can make the review session button appear at
the base of the session box.